Remove file list and input params from secrets build phase
They were unnecessary. Those values are translated in variables that the
build phase can read, but never did. The script defines its inputs and
outsides internally and that's appropriate because it needs logic to
decide which of the secrets to copy. If the script was selecting the
file based on an index in the array of inputs, it would be brittle
because the order might change without the script knowing about it.

Case in point. The file list didn't include the Jetpack secrets, but
the Jetpack target worked nevertheless.

On the other hand, it appears that the output list IS necessary. Without
it, I couldn't get the app to build with the Fastlane automation.

See:

- Build failing at
  https://app.circleci.com/pipelines/github/wordpress-mobile/WordPress-iOS/21768/workflows/b9ccaed2-4243-4491-9d59-c1ac5c606c66/jobs/49731
- Build passing* after reintroducing the output list with the file
  https://app.circleci.com/pipelines/github/wordpress-mobile/WordPress-iOS/21769/workflows/4cae0dbf-a99c-41a2-b01d-2610251a524f/jobs/49733

* It doesn't really pass, but it does fail in the expected way: with
TestFlight rejecting a build that was successfully uploaded but had an
already submitted build number.